cleaning up logs, crux instructions optional
